improve handling of cheats to handle things other than WRAM (closes #97)
note: mirroring detection for cheats doesn't work for two addresses in
the same 64kb bank i.e. the lowest 16 bits of both addresses must still
be the same

+++ bsnes/snes/cheat/cheat-inline.hpp
@@ -1,2 +1,2 @@
 bool Cheat::active() const { return cheat_enabled; }
-bool Cheat::exists(unsigned addr) const { return bitmask[addr >> 3] & 1 << (addr & 7); }
+bool Cheat::exists(uint16 addr) const { return bitmask[addr >> 3] & 1 << (addr & 7); }
